made sure kakasi dependent dicts were in assets (consolidated kakasi libs folder to use the kakasi crate)

+++ prototypes/libkakasi-cpp-test/build.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
+if [ ! -d build ]; then
+    mkdir build
+fi
 cd build
-rm CMakeCache.txt 
-cmake .. 
+if [ -f CMakeCache.txt ]; then
+    rm CMakeCache.txt
+fi
+
+# NOTE:  in case ninja is installed we'll make sure to force it to use Makefile instead of ninja.build
+cmake -G "Unix Makefiles" .. 
 make --trace
 find . -executable
